Add files via upload
The AI counselor, powered by Convai AI, provides personalized and accessible mental health support to users. It enables students to interact with an AI-powered virtual counselor, offering immediate and anonymous responses to their concerns. This innovative solution ensures that mental health services are approachable, timely, and tailored to individual needs, creating a supportive environment for users.

https://drive.google.com/file/d/1pYl1-7op72xtIA47yBXE28QKKzUgGbAD/view?usp=sharing

This is my drive link for the project.